chore: move protocolUnitTest, remove unnused timestampformat
